% Generated by roxygen2 (4.0.1): do not edit by hand
\name{colTrans}
\alias{colTrans}
\title{A pseudo transparency to a color}
\usage{
colTrans(col, f = 2)
}
\arguments{
\item{col}{A color}

\item{f}{The amount of pseudotransparency}
}
\value{
color
}
\description{
A pseudo transparency to a color
}
\author{
mg14
}

