% Generated by roxygen2 (4.0.1): do not edit by hand
\name{entropy}
\alias{entropy}
\title{Compute entropy (base e) from a vector of probabilities}
\usage{
entropy(p)
}
\arguments{
\item{p}{Vector of probabilities}
}
\value{
numeric
}
\description{
Given p, the entropy is computed as -sum(log(p^p)).
}
\author{
mg14
}

