\name{mergeLevels}
\alias{mergeLevels}
\title{Merge levels of a factor}
\usage{
  mergeLevels(f, mergeList)
}
\arguments{
  \item{f}{A factor}

  \item{mergeList}{A list of type list(newlevel1 =
  c(oldlevel1, oldlevel2), newlevel2=c(...),...)}
}
\value{
  A factor
}
\description{
  Merge levels of a factor
}
\author{
  mg14
}

