% Generated by roxygen2 (4.0.1): do not edit by hand
\name{numericalize}
\alias{numericalize}
\title{Function to convert factors to numeric}
\usage{
numericalize(x)
}
\arguments{
\item{x}{A vector or data.frame}
}
\value{
A data.frame with numeric vectors, or a factors (if x is non-numeric), or a single vector if x was a vector
}
\description{
Function to convert factors to numeric
}
\author{
mg14
}

