\name{rotatedLabel}
\alias{rotatedLabel}
\title{Plot a rotated axis label}
\usage{
  rotatedLabel(x0, y0, labels, pos = 1, cex = 1, srt = 45,
    ...)
}
\arguments{
  \item{x0}{The x position}

  \item{y0}{The y position}

  \item{labels}{The labels}

  \item{pos}{The positioning relative to the coordinate.
  Default = 1 (below).}

  \item{cex}{The character expansion factor. Default = 1.}

  \item{srt}{The degree of rotation. Default = 45.}

  \item{...}{}
}
\value{
  NULL
}
\description{
  Plot a rotated axis label
}
\author{
  mg14
}

