% Generated by roxygen2 (4.0.1): do not edit by hand
\name{stars}
\alias{stars}
\title{Stars plot}
\usage{
stars(x, full = TRUE, scale = TRUE, radius = TRUE,
  labels = dimnames(x)[[1L]], locations = NULL, nrow = NULL,
  ncol = NULL, len = 1, key.loc = NULL, key.labels = dimnames(x)[[2L]],
  key.xpd = TRUE, xlim = NULL, ylim = NULL, flip.labels = NULL,
  draw.segments = FALSE, col.segments = 1L:n.seg, col.stars = NA,
  col.lines = NA, axes = FALSE, frame.plot = axes, main = NULL,
  sub = NULL, xlab = "", ylab = "", cex = 0.8, lwd = 0.25,
  lty = par("lty"), xpd = FALSE, mar = pmin(par("mar"), 1.1 + c(2 * axes +
  (xlab != ""), 2 * axes + (ylab != ""), 1, 0)), add = FALSE, plot = TRUE,
  density = NA, init.angle = 0, ...)
}
\arguments{
\item{x}{data.frame or matrix to plot}
}
\description{
Modified from graphics::stars()
}
\author{
mg14
}

