% Generated by roxygen2 (4.0.1): do not edit by hand
\name{violinJitter}
\alias{violinJitter}
\title{Jitter by violins}
\usage{
violinJitter(x, magnitude = 1)
}
\arguments{
\item{x}{A vector of numbers}

\item{magnitude}{The overall magnitude of scatter}
}
\description{
Jitter by violins
}
\author{
mg14
}

