% Generated by roxygen2 (4.0.1): do not edit by hand
\name{violinJitterPlot}
\alias{violinJitterPlot}
\title{Violins plot with jittered data points}
\usage{
violinJitterPlot(y, x = factor(rep(1, length(y))), col = 1:nlevels(x),
  col.pty = colTrans(col), magnitude = 1, ...)
}
\arguments{
\item{y}{the values to be plotted}

\item{x}{the levels to be jittered}

\item{col}{the colors to be used for the outline}

\item{col.pty}{the colors used for the points}

\item{magnitude}{the magnitude of the violins}

\item{...}{additional parameters passed to plot()}
}
\description{
Violins plot with jittered data points
}
\examples{
f <- rep(1:5, each=50)
violinJitterPlot(rnorm(length(f), f), factor(f))
}
\author{
mg14
}

