% Generated by roxygen2 (4.0.1): do not edit by hand
\name{rotatedLabel}
\alias{rotatedLabel}
\title{Plot a rotated axis label}
\usage{
rotatedLabel(x0 = seq_along(labels), y0 = rep(par("usr")[3],
  length(labels)), labels, pos = 1, cex = 1, srt = 45, ...)
}
\arguments{
\item{x0}{The x position}

\item{y0}{The y position}

\item{labels}{The labels}

\item{pos}{The positioning relative to the coordinate. Default = 1 (below).}

\item{cex}{The character expansion factor. Default = 1.}

\item{srt}{The degree of rotation. Default = 45.}

\item{...}{}
}
\description{
Plot a rotated axis label
}
\author{
mg14
}

